+++ resources/views/layouts/includes/header.blade.php
@@ -37,7 +37,7 @@
                 <ul class="font-medium flex flex-col p-4 mt-4 border rounded-lg bg-gray-800 border-gray-700">
                     <li>
                         @if(Auth::check())
-                            <a href="#" class="block py-3 pl-3 pr-4 rounded hover:bg-gray-700 hover:text-blue-500 {{--@if(request()->routeIs('home')) text-blue-500 @endif--}}" aria-current="page">Profile</a>
+                            <a href="{{ route('profile') }}" class="block py-3 pl-3 pr-4 rounded hover:bg-gray-700 hover:text-blue-500 @if(request()->routeIs('profile')) text-blue-500 @endif" aria-current="page">Profile</a>
                             <form action="{{ route('logout') }}" method="POST" class="block py-3 pl-3 pr-4 rounded hover:bg-gray-700 hover:text-blue-500">
                                 @csrf
                                 <button type="submit" class="text-left w-full" aria-current="page">Log Out</button>